Splitting init and sync into separate commands
